requirements.txt
altair==5.5.0
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-genai==1.21.1
idna==3.10
matplotlib==3.10.3
munkres==1.1.4
narwhals==1.44.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
plotly==6.2.0
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.15.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyPDF2==3.0.1
pypdfium2==4.30.1
rsa==4.9.1
seaborn==0.13.2
setuptools==78.1.1
smmap==5.0.2
streamlit==1.46.1
streamlit-card==1.0.2
tenacity==8.5.0
toml==0.10.2    
typing-inspection==0.4.1
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1